feat: new file for single ft quote_removal
